+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # BusInjectionModel Change log
 
+## dev
+- add classic DC model `build_bim_polar(m::JuMP.AbstractModel, net::Network{SinglePhase}, ::Val{DC})`
+
 ## v0.1.2
 - add fixed point linear model, multi and single phase
     - leverage `CommonOPF.BusTerminal` as model indices s.t. do not have to implement single and